move feedback into index
